split account.js and made login and register seperate components
